requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bleach==5.0.1
DateTime==4.7
Django==4.1.1
django-appconf==1.0.5
django-bleach==3.0.0
django-crispy-forms==1.14.0
django-imagekit==4.1.0
django-tagging==0.5.0
djangorestframework==3.13.1
gunicorn==20.1.0
html-sanitizer==1.9.3
lxml==4.9.1
pilkit==2.0
Pillow==9.2.0
pytz==2022.2.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tinycss2==1.1.1
tzdata==2022.2
webencodings==0.5.1
whitenoise==6.2.0
zope.interface==5.4.0
